+++ src/main/java/com/svam/repository/AHTRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.svam.models.AgencyHearingTime;
 
-public interface AHTRepository extends CrudRepository<AgencyHearingTime, Long>{
+public interface AHTRepository extends CrudRepository<AgencyHearingTime, Integer>{
 
 	List<AgencyHearingTime>  findByAgencyName(String agencyName);
 }
